feat: Parallelize feature building and TSCV (feature)
This commit introduces parallelization of feature building and
Time Series Cross-Validation (TSCV) using Ray. It includes:

- Parallelized feature engineering using `build_features_task`.
- Parallelized TSCV fold computation using
  `compute_single_time_series_cv_fold_task`.
- Updated `validator.py` to include a task for single fold
  computation.
- Added `features/builder.py` for feature building via Ray.
- Updated data exploration to initialize Ray.
- Modified const.py to include dataset index reset.
- Fixed a bug in forward runner

+++ src/components/backpropagation/core/forward_runner.py
@@ -59,7 +59,7 @@ def compute_forward(
         # Move batch to device
         x_features = move_to_device(x_features, device)
         x_keys = move_to_device(x_keys, device)
-        y_key = move_to_device(y_key, device)
+        y_key = move_to_device(y_key, device).long()
 
         # Compute forward pass
         outputs = model(x_features, x_keys)
